requirement.txt
appdirs==1.4.3
autopep8==1.3.1
click==6.7
decorator==4.0.11
Django==1.11
facebook-scribe-py3==0.1.1
flake8==3.3.0
Flask==0.12.1
gunicorn==19.7.1
image==1.5.5
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
packaging==16.8
pdb==0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-gnupg==0.4.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.13.0
selenium==3.3.1
simplegeneric==0.8.1
six==1.10.0
thrift==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.1
